Estructura
- Modelos basicos de model, category y enpoints

+++ src/entity/Model.ts
@@ -10,7 +10,7 @@ export class Model {
     @Column()
     key: string;
 
-    @Column("float")
+    @Column("int")
     quantity: number;
 
     @Column()
